refactor(ffi): move methods from `RoomListItem` to `Room`
Only code motion. Most methods already existed on the `Room` impl block.
